requirements.txt
telethon>=1.36.0
python-dotenv>=1.0.0
pywin32>=306
pyautogui>=0.9.54
Pillow>=10.0.0
pystray>=0.19.5
